+++ examples/ai-core/src/generate-image/google-vertex.ts
@@ -1,7 +1,10 @@
-import 'dotenv/config';
-import { vertex } from '@ai-sdk/google-vertex';
+import {
+  GoogleVertexImageProviderOptions,
+  vertex,
+} from '@ai-sdk/google-vertex';
 import { experimental_generateImage as generateImage } from 'ai';
-import fs from 'fs';
+import 'dotenv/config';
+import { presentImages } from '../lib/present-image';
 
 async function main() {
   const { image } = await generateImage({
@@ -10,15 +13,12 @@ async function main() {
     aspectRatio: '1:1',
     providerOptions: {
       vertex: {
-        // https://cloud.google.com/vertex-ai/generative-ai/docs/model-reference/imagen-api#parameter_list
         addWatermark: false,
-      },
+      } satisfies GoogleVertexImageProviderOptions,
     },
   });
 
-  const filename = `image-${Date.now()}.png`;
-  fs.writeFileSync(filename, image.uint8Array);
-  console.log(`Image saved to ${filename}`);
+  await presentImages([image]);
 }
 
 main().catch(console.error);